refactor(finance_account): use Decimal for balance delta calculations
Replace float conversion with Decimal for delta_absolute calculation in balance trend service. Using Decimal ensures proper precision in financial calculations and prevents floating-point rounding errors that can accumulate in monetary calculations.

+++ hasta_la_vista_money/finance_account/services/balance_trend_service.py
@@ -96,7 +96,7 @@ class BalanceTrendService:
         )
 
         # Calculate delta
-        delta_absolute = float(current_balance - period_start_balance)
+        delta_absolute = Decimal(current_balance - period_start_balance)
         delta_percent = (
             (delta_absolute / period_start_balance * 100)
             if period_start_balance != 0
